Use python2 instead of python (#57)
This tool will only work with Python 2, but for quite a few systems
/usr/bin/python is Python 3.

This should make it work well for Python 3 systems too, or make it fail
with a clear error if Python 2 isn't installed. As far as I know all
systems have been installing `python2` for as long as I can remember, so
this shouldn't break anything.